[build.webkit.org] [GTK] Remove 'additionalArguments' in GTK3 bot config https://bugs.webkit.org/show_bug.cgi?id=278853
Reviewed by Carlos Alberto Lopez Perez.

The GTK3 post-commit bot is a bot-and-test bot. It seems the parameters
in 'additonalArguments' are passed also to test steps such as
'layout-test', 'dashboard-test' and 'api-tests', causing the tests to fail
since the arguments passed are unrecognized by the scripts in charge of
running those tests.

* Tools/CISupport/build-webkit-org/config.json:

Canonical link: https://commits.webkit.org/283551@main