Do more validation of IPC::Decoder's destinationID https://bugs.webkit.org/show_bug.cgi?id=279574 rdar://135525328
Reviewed by Darin Adler.

Do more validation of IPC::Decoder's destinationID. We have to keep allowing 0
since it is a valid destinationID. However, we can reject -1 which is the
HashTable's deleted value.

This is important since we sometimes constructor ObjectIdentifiers from the
decoder's destinationID and these ObjectIdentifiers may be used to look up
in a HashMap.

283525@main also added more validation of the raw identifier when constructing
an ObjectIdentifier which will further protect us.

* Source/WebKit/Platform/IPC/Decoder.cpp:

Canonical link: https://commits.webkit.org/280938.329@safari-7619-branch