Remove Quirk triggerOptionalStorageAccessQuirk for domains not part of Kinja anymore https://bugs.webkit.org/show_bug.cgi?id=279581 rdar://135854408
Reviewed by Brent Fulgham.

kinjaQuirks defines a list of domains to be used by the Quirk
triggerOptionalStorageAccessQuirk. This list is then used later in
addition to isKinjaLoginAvatarElement (another Quirk)

if (kinjaQuirks.get().contains(domain)
    && isKinjaLoginAvatarElement(element)) { … }

Some of these domains are not part anymore of Kinja Login system. They
can be removed.

The list of domains to remove is:

* avclub.com
* deadspin.com
* jezebel.com
* lifehacker.com
* thetakeout.com
* theonion.com

It would be good to check for the remaining domains if the quirk is
needed at all. The markup defined in isKinjaLoginAvatarElement still
exists.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::triggerOptionalStorageAccessQuirk const):

Canonical link: https://commits.webkit.org/283581@main