Remove Quirk isStorageAccessQuirkDomainAndElement for gizmodo.com https://bugs.webkit.org/show_bug.cgi?id=279636 rdar://135927150
Reviewed by Sihui Liu.

gizmodo is not part anymore of the Kinja login system.
They also have removed the icon at the top right which was initially
there. On some articles, there is the possibility to comment as Guest
or logged in which is powered by a different tracking company. After
testing with both Site Specific Hacks Enabled and Disabled, we can
conclude the login for comments is working.

The quirk is then not necessary anymore.

* Source/WebCore/page/Quirks.cpp:
(WebCore::isStorageAccessQuirkDomainAndElement):
(WebCore::elementHasClassInClosestAncestors): Deleted.
* Source/WebCore/platform/network/NetworkStorageSession.cpp:
(WebCore::NetworkStorageSession::storageAccessQuirks):

Canonical link: https://commits.webkit.org/283600@main