Improve HashTable::checkKey() https://bugs.webkit.org/show_bug.cgi?id=279686
Reviewed by Darin Adler.

update HashTable::checkKey() to rely on KeyTraits::isEmptyValue() & KeyTraits::isDeletedValue()
whenever possible. This is better for performance as it avoids constructing those values to do
an `==` comparison. This also enables us to validate the key in more cases. We would previously
not do ANY validation of the key if `HashFunctions::safeToCompareToEmptyOrDeleted` was false.

* Source/WTF/wtf/HashTable.h:
(WTF::KeyTraits>::checkKey):

Canonical link: https://commits.webkit.org/283670@main