Table layout inconsistency with <td withd="100%"> https://bugs.webkit.org/show_bug.cgi?id=279450
Reviewed by Antti Koivisto.

This is about removing an old quirk (3673@main) which was added to make excite.com look better by
not letting percent based cell bloat in some cases (nested table, auto width).
While it might visually be the preferred approach, it hurts interoperability as neither Blink nor Gecko adopted it.
It is also inconsistent as when the container of such table content is not an other table
but a min-content sized block container we stop applying this quirk and let the cell bloat.

(LayoutTests are progressions and now we match Blink)

* LayoutTests/fast/table/stretchy-table-with-percent-cell-expected.html: Added.
* LayoutTests/fast/table/stretchy-table-with-percent-cell.html: Added.
* Source/WebCore/rendering/AutoTableLayout.cpp:
(WebCore::AutoTableLayout::computeIntrinsicLogicalWidths):
(WebCore::shouldScaleColumnsForSelf): Deleted.

Canonical link: https://commits.webkit.org/283711@main