Remove Quirk shouldDisableNavigatorStandaloneQuirk for oracle.com https://bugs.webkit.org/show_bug.cgi?id=279670 rdar://135948711
Reviewed by Brent Fulgham.

navigator.standalone had been disabled through a Quirk for oracle.com.
The site was using it to identify exclusively the iPad and iOs with a
combination of maxTouchPoints, this was breaking the charts in Oracle.

There's no traces anymore of standalone in the oracle.com code
and the charts are working equally with or without site specific hacks.

* Source/WebCore/page/Navigator.idl:
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldDisableNavigatorStandaloneQuirk const): Deleted.
* Source/WebCore/page/Quirks.h:

Canonical link: https://commits.webkit.org/283728@main