[JSC] replace m_frozenValuesAreFinalized check with safepoint check https://bugs.webkit.org/show_bug.cgi?id=279824 rdar://136150059
Reviewed by Yusuke Suzuki.

Now that the plan knows whether the compiler is inside a GC safepoint
or not we can directly check that rather than having separate
state for this assertion. That will be more robust if we allow more
compilation to occur within a GC safepoint. Also, the name
m_frozenValuesAreFinalized was misleading since it is not a "final"
state.

* Source/JavaScriptCore/dfg/DFGGraph.cpp:
(JSC::DFG::Graph::freeze):
* Source/JavaScriptCore/dfg/DFGGraph.h:
* Source/JavaScriptCore/ftl/FTLCompile.cpp:
(JSC::FTL::compile):
* Source/JavaScriptCore/jit/JITPlan.cpp:
(JSC::JITPlan::isInSafepoint const):
* Source/JavaScriptCore/jit/JITPlan.h:

Canonical link: https://commits.webkit.org/283789@main