[GStreamer][DMABuf] Disable DMABuf video sink by default https://bugs.webkit.org/show_bug.cgi?id=279819
Reviewed by Xabier Rodriguez-Calvar.

In its current form this sink and the associated rendering logic in the TextureMapper are prone to
memory leaks and increased file descriptors usage. There is a plan to rework this but it was too
late for the 2.46 release.

* Source/WebCore/platform/graphics/gstreamer/DMABufVideoSinkGStreamer.cpp:
(webKitDMABufVideoSinkIsEnabled):
* Tools/Scripts/webkitpy/port/glib.py:
(GLibPort.setup_environ_for_server):

Canonical link: https://commits.webkit.org/283833@main